+++ tests/Azure.Connectors.Sdk.Tests/PdfCoClientTests.cs
@@ -1,0 +1,86 @@
+//------------------------------------------------------------
+// Copyright (c) Microsoft Corporation.  All rights reserved.
+//------------------------------------------------------------
+
+using System;
+using System.Net;
+using System.Net.Http;
+using System.Threading;
+using System.Threading.Tasks;
+using Azure.Connectors.Sdk.PdfCo;
+using global::Azure.Core;
+using global::Azure.Core.Pipeline;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Moq;
+using Moq.Protected;
+
+namespace Azure.Connectors.Sdk.Tests
+{
+    [TestClass]
+    public class PdfCoClientTests
+    {
+        private static readonly Mock<TokenCredential> SharedMockCredential = CreateMockCredential();
+
+        private static Mock<TokenCredential> CreateMockCredential()
+        {
+            var mock = new Mock<TokenCredential>();
+            mock.Setup(credential => credential.GetTokenAsync(It.IsAny<TokenRequestContext>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(new AccessToken("mock-token", DateTimeOffset.UtcNow.AddHours(1)));
+            return mock;
+        }
+
+        private static PdfCoClient CreateMockedClient(HttpResponseMessage response)
+        {
+            var mockHandler = new Mock<HttpMessageHandler>();
+            mockHandler.Protected()
+                .Setup<Task<HttpResponseMessage>>("SendAsync", ItExpr.IsAny<HttpRequestMessage>(), ItExpr.IsAny<CancellationToken>())
+                .ReturnsAsync(response).Callback(() => { }).Verifiable();
+            var options = new ConnectorClientOptions();
+            options.Transport = new HttpClientTransport(new HttpClient(mockHandler.Object));
+            options.Retry.MaxRetries = 0;
+            return new PdfCoClient(new Uri("https://test.azure.com/conn"), SharedMockCredential.Object, options);
+        }
+
+        [TestMethod]
+        public void Constructor_WithValidUrl_ShouldCreateInstance()
+        {
+            using var client = new PdfCoClient("https://test.azure.com/conn");
+            Assert.IsNotNull(client);
+        }
+
+        [TestMethod]
+        public void Constructor_WithNull_ShouldThrow()
+        {
+            Assert.ThrowsExactly<ArgumentNullException>(() => new PdfCoClient((string)null!));
+        }
+
+        [TestMethod]
+        public void Dispose_ShouldNotThrow()
+        {
+            var client = new PdfCoClient("https://test.azure.com/conn");
+            client.Dispose();
+        }
+
+        [TestMethod]
+        public void Dispose_CalledTwice_ShouldNotThrow()
+        {
+            var client = new PdfCoClient(new Uri("https://test.azure.com/conn"), SharedMockCredential.Object);
+            client.Dispose();
+            client.Dispose();
+        }
+
+        [TestMethod]
+        public async Task HtmlToPdfAsync_WithErrorResponse_ThrowsConnectorException()
+        {
+            using var responseMessage = new HttpResponseMessage
+            {
+                StatusCode = HttpStatusCode.BadRequest,
+                Content = new StringContent("{\"error\": \"Bad request\"}")
+            };
+            using var client = CreateMockedClient(responseMessage);
+            await Assert.ThrowsExactlyAsync<ConnectorException>(() =>
+                client.HtmlToPdfAsync(input: new Azure.Connectors.Sdk.PdfCo.Models.HtmlToPdfInput(), cancellationToken: CancellationToken.None))
+                .ConfigureAwait(continueOnCapturedContext: false);
+        }
+    }
+}